+++ MotelRoomManagement/MotelRoomManagement/frmChangePass.cs
@@ -7,6 +7,9 @@ using System.Linq;
 using System.Text;
 using System.Windows.Forms;
 
+using MRBUS;
+using MRDTO;
+
 namespace MotelRoomManagement
 {
     public partial class frmChangePass : Form
@@ -37,15 +40,32 @@ namespace MotelRoomManagement
             oldP = txtOldPass.Text;
             newP = txtNewPass.Text;
             rewrite = txtRewrite.Text;
-            if (pass == oldP)
+            if (oldP == "" || newP == "" || rewrite == "")
             {
-                if (newP == rewrite)
-                    MessageBox.Show("OK");
-                else
-                    MessageBox.Show("Vui lòng nhập lại mật khẩu mới", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    MessageBox.Show("Vui lòng điền đủ thông tin", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+
             }
             else
-                MessageBox.Show("Mật khẩu không đúng", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            {
+                if (pass == oldP)
+                {
+                    if (newP == rewrite)
+                    {
+                        int i = new DangNhapBUS().ChangePass(id, newP);
+                        MessageBox.Show("Thay đổi thành công");
+                        Close();
+                    }
+                    else
+                        MessageBox.Show("Vui lòng nhập lại mật khẩu mới", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                }
+                else
+                    MessageBox.Show("Mật khẩu không đúng", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            }
+        }
+
+        private void btCancel_Click(object sender, EventArgs e)
+        {
+            Close();
         }
     }
 }
